fix system 2/2 issue

+++ src/BE/Services/Conversations/Dtos/ModelConfig.cs
@@ -9,7 +9,7 @@ public record ModelConfig
     public required string Prompt { get; init; }
 
     [JsonPropertyName("temperature")]
-    public required double Temperature { get; init; }
+    public required float Temperature { get; init; }
 
     [JsonPropertyName("enableSearch"), JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
     public bool? EnableSearch { get; init; }
